Split the dandelion functions out of net/net_processing.
